+++ CHANGES.md
@@ -11,6 +11,7 @@ NEXT MILESTONE
 - Add `Solver.tableSampling` resolution policy which returns a stream of `Solution` randomly sampled.
 - Extend some black-box strategies to `SetVar` and `GraphVar`
 - Add `union(SetVar unionSet, SetVar indices, SetVar[] sets)` constraint
+- Add `increasing` and `decreasing` constraints
  
 ### Deprecated API (to be removed in next release):